checkpatch: check signoff when reading stdin
Before, signoff was not checked if the filename is '-', indicating
reading the patch from stdin.  This causes commands such as below not to
warn about a missing signoff.

  git show --pretty=email | scripts/checkpatch.pl -

As a workaround, the command could be modified to refer to stdin by a
name other than '-'.  The workaround is not an elegant solution, because
elsewhere checkpatch uses the fact that filename equals '-' is special
for stdin, such as setting '$vname' to 'Your patch'.

  git show --pretty=email | scripts/checkpatch.pl /dev/stdin

This change causes checkpatch to check for a missing signoff line, even
if the filename is '-', as in the first variation of the command.

Signed-off-by: Allen Hubbe <allenbh@gmail.com>